BustAMovePanel::OnBeat: 97.69 -> 98.3, three real bugs plus the half-applied temp lever
Three behavioural fixes, all read off the target listing:
- mRecorder->SetVal44(mBeatCount) stored the beat count into
  FreestyleMoveRecorder+0x44 (mSkeletonIndex); the target stores to +0x40
  (idx 533). Named the field mBeatIndex and added SetBeatIndex(); RhythmBattle's
  SetVal44 (which really is the skeleton index) is untouched.
- MILO_ASSERT(mReps == 0, 0x328) tested mRepsRemaining; the target loads
  +0x44 = mBeatCount (idx 1205). The original's name for mBeatCount was mReps.
- mPlayerScoreLeft/Right = 0 stored an int; the target does stfs of 0.0f
  (idx 3002-3003). Members are now float (the other user already cast them),
  and that alone flipped the f29/f30/f31 constant assignment to match.

Match-only:
- The remaining mFlashcardLabels.push_back(gNullStr) sites (implicit ctor)
  became push_back(Symbol(gNullStr)) so the ctor return is forwarded (mr r6,r3)
  instead of re-materialised (addi r6,r31,slot): 13 replace/insert rows.
- sideStr in PlayCountIn is const char*, so MakeString<const char*> is
  instantiated as in the target (one name_check row).

+++ src/lazer/game/BustAMovePanel.h
@@ -92,8 +92,8 @@ private:
     int mMoveIndex; // 0x84 - current freestyle move slot (0-3)
     int unk88; // 0x88
     int unk8c; // 0x8c
-    int mPlayerScoreLeft; // 0x90 - reinterpreted as float for final sequence scoring
-    int mPlayerScoreRight; // 0x94 - reinterpreted as float for final sequence scoring
+    float mPlayerScoreLeft; // 0x90 - final sequence score, left player
+    float mPlayerScoreRight; // 0x94 - final sequence score, right player
     RndDir *mBAMColumns[kNumSkeletonSides]; // 0x98
     int mCreatorSide; // 0xa0 - side of the move creator (SkeletonSide)
     DancerSkeleton mCapturePoses[3]; // 0xa4 - skeleton poses for flashcard capture
